pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeindexer"
version = "0.2.0"
description = "Code indexing and search system with vector and graph storage"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "CodeIndexer Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "numpy>=1.20.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
    "neo4j>=5.5.0",
    "tree-sitter>=0.20.0",
    "gitpython>=3.1.30",
    "tqdm>=4.64.0",
    "loguru>=0.6.0",
    "tenacity>=8.1.0",
]

[project.optional-dependencies]
semantic = [
    "google-adk>=0.5.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0"
]
vector = [
    "pymilvus>=2.4.0",
    "sentence-transformers>=2.2.0"
]
qdrant = [
    "qdrant-client>=1.7.0"
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0"
]

[project.urls]
"Homepage" = "https://github.com/username/codeindexer"
"Bug Tracker" = "https://github.com/username/codeindexer/issues"

[tool.setuptools]
packages = ["code_indexer"]

[project.scripts]
codeindexer-ingest = "code_indexer.ingestion.cli:main"
codeindexer-semantic = "code_indexer.semantic.api:main"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100